Hide settings menus by default (TRAX patch 1)
This patch disable the following menus:
- Settings
- Tune
- Calibration
- Statistics
- Fail stats
- Support

English is the only supported language for this support - other languages are not built by default.

This is to ensure that the printers are not mishandled by TRAX members.
Only board members are allowed to change these settings using an obfuscated switch.
A consequence of this patch is that the logic analyzer can not be used together with the MK3.

See https://github.com/ChalmersTRAX/Prusa-Firmware/wiki/Locked-settings-mode
for more information.